requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.80
botocore==1.31.80
cachelib==0.12.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.3.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
Flask-Session==0.6.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.0.0
lxml==4.9.3
MarkupSafe==2.1.3
mongoengine==0.27.0
mongomock==3.23.0
moto==2.3.1
nltk==3.6.7
pycparser==2.21
pymongo==4.6.2
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
responses==0.24.0
s3transfer==0.7.0
sentinels==1.0.0
six==1.16.0
tqdm==4.66.1
urllib3==1.26.18
Werkzeug==2.3.7
xlrd==1.2.0
xmldiff==2.4
xmltodict==0.13.0
zipp==3.16.2
